solve 2 problems

+++ src/course3/fibonacci.java
@@ -5,11 +5,11 @@ import java.util.Scanner;
 public class fibonacci {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        long n = sc.nextLong();
         System.out.println(fibonacci(n));
     }
 
-    private static long fibonacci(int n) {
+    private static long fibonacci(long n) {
         if (n <= 1) return n;
         return fibonacci(n - 1) + fibonacci(n - 2);
     }
